feat: useTimer 커스텀 훅 작성
