+++ src/components/common/AuthTooltip/AuthTooltip.tsx
@@ -1,35 +1,15 @@
 import React, { useEffect, useState } from "react";
 import RefreshButton from "../../../assets/svg/RefreshButton";
+import useTimer from "../../../hooks/useTimer";
 
 interface AuthTooltipProps {
   isAuth: boolean;
 }
 
 const AuthTooltip = ({ isAuth }: AuthTooltipProps) => {
-  const MINUTES_IN_MS = 60 * 60 * 1000 - 1000;
-  const INTERVAL = 1000;
-  const [leftTime, setLeftTime] = useState<number>(MINUTES_IN_MS);
+  const initialTime = 60 * 60 * 1000 - 1000;
 
-  const minutes = String(Math.floor((leftTime / (1000 * 60)) % 60)).padStart(
-    2,
-    "0",
-  );
-  const second = String(Math.floor((leftTime / 1000) % 60)).padStart(2, "0");
-
-  useEffect(() => {
-    const timer = setInterval(() => {
-      setLeftTime((prevTime) => prevTime - INTERVAL);
-    }, INTERVAL);
-
-    if (leftTime <= 0) {
-      clearInterval(timer);
-      // 종료 로직
-    }
-
-    return () => {
-      clearInterval(timer);
-    };
-  }, [leftTime]);
+  const { reset, minutes, second } = useTimer(initialTime);
 
   return (
     <>
@@ -45,9 +25,7 @@ const AuthTooltip = ({ isAuth }: AuthTooltipProps) => {
             <span className="text-center font-kia-signature-bold text-body-2-bold">
               {minutes} : {second}
             </span>
-            <RefreshButton
-              onClick={() => setLeftTime(MINUTES_IN_MS)}
-            ></RefreshButton>
+            <RefreshButton onClick={reset}></RefreshButton>
           </div>
         </div>
       ) : (
